% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStatistics.R
\name{addStatistics.giotto}
\alias{addStatistics.giotto}
\title{Add statistics to regular giotto object}
\usage{
\method{addStatistics}{giotto}(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  stats = c("feature", "cell", "area"),
  expression_values = c("normalized", "scaled", "custom"),
  detection_threshold = 0,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{stats}{character. What statistics to add.
default = c("cell", "feature") See details}

\item{expression_values}{expression values to use}

\item{detection_threshold}{detection threshold to consider a feature detected}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{verbose}{be verbose}
}
\value{
giotto object if return_gobject = TRUE, else a list with results
}
\description{
Default method that delegates to the original Giotto implementation
}
