% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_giotto.R
\name{as_giotto}
\alias{as_giotto}
\title{Convert GiottoDB Object to giotto}
\usage{
as_giotto(x, verbose = TRUE)
}
\arguments{
\item{x}{A \code{\link{GiottoDB}} object}

\item{verbose}{Whether to print progress messages}
}
\value{
A \code{\link[GiottoClass:giotto-class]{GiottoClass::giotto}} object
}
\description{
A convenience function that coerces a \code{GiottoDB} object to an in-memory
\code{giotto} object. Expression matrices are converted to in-memory matrices
using dbverse coercions:
\itemize{
\item \code{dbSparseMatrix}/\code{dbMatrix} -> \code{dgCMatrix}
\item \code{dbDenseMatrix} -> base \code{matrix}
}

Spatial data stored as \code{dbSpatial} are converted to \code{terra::SpatVector}.
}
\examples{
\dontrun{
library(GiottoDB)
library(GiottoData)
library(duckdb)
library(DBI)

g <- GiottoData::loadGiottoMini("visium")
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = ":memory:")
gdb <- as_giottodb(g, con = con)

g_inmem <- as_giotto(gdb)
DBI::dbDisconnect(con, shutdown = TRUE)
}
}
\concept{Object creation}
