% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_giottodb.R
\name{as_giottodb}
\alias{as_giottodb}
\title{Convert giotto Object to GiottoDB}
\usage{
as_giottodb(
  x,
  con = NULL,
  db_path = NULL,
  prefix = "gdb_",
  overwrite = FALSE,
  verbose = TRUE,
  temporary = TRUE
)
}
\arguments{
\item{x}{A giotto object}

\item{con}{A \link[DBI:DBI-package]{DBI::DBI} connection from a \link[duckdb:duckdb]{duckdb::duckdb} connection object}

\item{db_path}{Path to the database file if creating a new persistent connection}

\item{prefix}{A string prefix to add to database table names}

\item{overwrite}{Whether to overwrite existing tables}

\item{verbose}{Whether to print progress messages}

\item{temporary}{if TRUE (default), will create a temporary table that is
local to this connection and will be automatically deleted when con expires}
}
\value{
A \code{\link{GiottoDB}} object
}
\description{
A convenience function that coerces a giotto object to a GiottoDB object
with a database backend. Expression matrices are converted to dbMatrix objects
and spatial data are converted to dbSpatial objects.
}
\examples{
\dontrun{
library(GiottoDB)
library(dbSpatial)

# Create connection
con <- dbConnect(duckdb(), dbdir = ":memory:")
loadSpatial(con)

# Convert Giotto object
my_giotto_db <- as_giottodb(my_giotto, con = con)

# Don't forget to close the connection when done
dbDisconnect(con, shutdown = TRUE)
}
}
