% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHVF_S3.R
\name{calculateHVF}
\alias{calculateHVF}
\alias{calculateHVF.GiottoDB}
\alias{calculateHVF.giotto}
\title{Calculate Highly Variable Features for GiottoDB}
\usage{
calculateHVF(gobject, ...)

\method{calculateHVF}{GiottoDB}(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = "normalized",
  method = c("cov_groups", "cov_loess", "var_p_resid"),
  reverse_log_scale = FALSE,
  logbase = 2,
  expression_threshold = 0,
  nr_expression_groups = 20,
  zscore_threshold = 1.5,
  HVFname = "hvf",
  difference_in_cov = 0.1,
  var_threshold = 1.5,
  var_number = NULL,
  random_subset = NULL,
  set_seed = TRUE,
  seed_number = 1234,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "HVFplot",
  return_gobject = TRUE,
  calc_gini = TRUE,
  verbose = TRUE
)

\method{calculateHVF}{giotto}(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  method = c("cov_groups", "cov_loess", "var_p_resid"),
  reverse_log_scale = FALSE,
  logbase = 2,
  expression_threshold = 0,
  nr_expression_groups = 20,
  zscore_threshold = 1.5,
  HVFname = "hvf",
  difference_in_cov = 0.1,
  var_threshold = 1.5,
  var_number = NULL,
  random_subset = NULL,
  set_seed = TRUE,
  seed_number = 1234,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "HVFplot",
  return_gobject = TRUE,
  calc_gini = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use}

\item{method}{method to calculate highly variable features}

\item{reverse_log_scale}{reverse log-scale of expression values
(default = FALSE)}

\item{logbase}{if \code{reverse_log_scale} is TRUE, which log base was used?}

\item{expression_threshold}{expression threshold to consider a gene detected}

\item{nr_expression_groups}{(cov_groups) number of expression groups for
cov_groups}

\item{zscore_threshold}{(cov_groups) zscore to select hvg for cov_groups}

\item{HVFname}{name for highly variable features in cell metadata}

\item{difference_in_cov}{(cov_loess) minimum difference in coefficient of
variance required}

\item{var_threshold}{(var_p_resid) variance threshold for features for
var_p_resid method}

\item{var_number}{(var_p_resid) number of top variance features for
var_p_resid method}

\item{random_subset}{random subset to perform HVF detection on.
Passing \code{NULL} runs HVF on all cells.}

\item{set_seed}{logical. whether to set a seed when random_subset is used}

\item{seed_number}{seed number to use when random_subset is used}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object (overridden by \code{return_gobject})}

\item{save_plot}{logical. directly save the plot}

\item{save_param}{list of saving parameters from
\code{\link[GiottoVisuals:plot_save]{GiottoVisuals::all_plots_save_function()}}}

\item{default_save_name}{default save name for saving, don't change, change
save_name in save_param}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{calc_gini}{logical. Whether to calculate Gini index for each feature.
Set to FALSE for performance with large datasets or dbMatrix objects.}

\item{verbose}{be verbose}
}
\description{
S3 method that prevents accessing "scaled" expression to avoid materialization.
}
\concept{Expression processing}
