% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOverlap.R
\name{calculateOverlap}
\alias{calculateOverlap}
\title{Calculate overlaps with dbSpatial-backed objects}
\usage{
\S4method{calculateOverlap}{dbSpatial,dbSpatial}(
  x,
  y,
  poly_subset_ids = NULL,
  feat_subset_column = NULL,
  feat_subset_values = NULL,
  feat_count_column = NULL,
  feat_subset_ids = NULL,
  count_info_column = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Object with spatial annotations: \code{giottoPolygon}, or \code{SpatVector}
polygons. Can also be a \code{giotto} object}

\item{y}{Object with features to overlap: \code{giottoPoints}, \code{giottoLargeImage},
\code{SpatVector} points or \code{SpatRaster}}

\item{poly_subset_ids}{character vector. (optional) Specific poly_IDs to use}

\item{feat_subset_column}{character. (optional) feature info attribute to
subset feature points on when performing overlap calculation.}

\item{feat_subset_values}{(optional) values matched against
in \code{feat_subset_column} in order to subset feature points when performing
overlap calculation.}

\item{feat_count_column}{character. (optional) column with count information.
Useful in cases when more than one detection is reported per point. If a
column called "count" is present in the feature points data, it will be
automatically selected.}

\item{feat_subset_ids}{deprecated. Use \code{feat_subset_values} instead.}

\item{count_info_column}{deprecated. Use \code{feat_count_column} instead.}

\item{verbose}{be verbose}

\item{...}{additional params to pass to methods.}
}
\value{
A dbSpatial object containing point features joined to overlapping
polygon metadata.
}
\description{
Compute overlaps between dbSpatial-backed polygon and point features while
keeping the result database-backed.
}
\concept{Subcellular workflow}
