% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{cast}
\alias{cast}
\alias{castNumeric,dbData,character-method}
\alias{castCharacter,dbData,character-method}
\alias{castLogical,dbData,character-method}
\alias{castNumeric,ANY,character-method}
\alias{castCharacter,ANY,character-method}
\alias{castLogical,ANY,character-method}
\title{Cast a column to a different type}
\usage{
\S4method{castNumeric}{dbData,character}(x, col, ...)

\S4method{castCharacter}{dbData,character}(x, col, ...)

\S4method{castLogical}{dbData,character}(x, col, ...)

\S4method{castNumeric}{ANY,character}(x, col, ...)

\S4method{castCharacter}{ANY,character}(x, col, ...)

\S4method{castLogical}{ANY,character}(x, col, ...)
}
\arguments{
\item{x}{GiottoDB data object}

\item{col}{column to cast}

\item{...}{additional params to pass}
}
\description{
Converts a column after first checking the column data type. Does
nothing if the column is already of that type
This precaution is to avoid truncation of values.
}
