% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGiottoPoints.R
\name{create_giotto_points_object_db}
\alias{create_giotto_points_object_db}
\title{Create giotto points object with dbSpatial}
\usage{
create_giotto_points_object_db(
  feat_type = "rna",
  dbSpatial = NULL,
  networks = NULL,
  unique_IDs = NULL
)
}
\arguments{
\item{feat_type}{feature type}

\item{dbSpatial}{dbSpatial object containing point data}

\item{networks}{(optional) feature network object}

\item{unique_IDs}{(optional) unique IDs for cacheing}
}
\value{
giotto_points_object
}
\description{
Create giotto points object with dbSpatial
}
\keyword{internal}
