% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGiottoPolygon.R
\name{create_giotto_polygon_object_db}
\alias{create_giotto_polygon_object_db}
\title{Create giotto polygon object with dbSpatial}
\usage{
create_giotto_polygon_object_db(
  name = "polygons",
  dbSpatial = NULL,
  unique_IDs = NULL,
  calc_centroids = FALSE
)
}
\arguments{
\item{name}{name for the polygon object}

\item{dbSpatial}{dbSpatial object containing polygon data}

\item{unique_IDs}{(optional) unique IDs for cacheing}

\item{calc_centroids}{logical. Whether to calculate centroids}
}
\value{
giotto_polygon_object
}
\description{
Create giotto polygon object with dbSpatial
}
\keyword{internal}
