% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatProxy.R
\name{createDBPointsProxy}
\alias{createDBPointsProxy}
\title{Create a framework for chunked processing with terra points SpatVectors}
\usage{
createDBPointsProxy(
  SpatVector,
  remote_name = "pnt_test",
  db_path = ":temp:",
  xy_col = c("x", "y"),
  extent = NULL,
  overwrite = FALSE,
  chunk_size = 10000L,
  callback = NULL,
  custom_table_fields = NULL,
  ...
)
}
\description{
Create an S4 dbPointsProxy object
}
\details{
Information is only read into the database during this process. Based
on the \code{remote_name} and \code{db_path} a lazy connection is then made
downstream during dbData intialization and appended to the object. If the data
already exists within the database backend then it is entirely permissible to
omit the \code{SpatVector} param.
}
