% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGiottoPoints.R
\name{createGiottoPoints}
\alias{createGiottoPoints}
\alias{createGiottoPoints,dbSpatial-method}
\title{Create GiottoPoints object using dbSpatial}
\usage{
\S4method{createGiottoPoints}{dbSpatial}(
  x,
  feat_type = "rna",
  verbose = TRUE,
  split_keyword = NULL,
  unique_IDs = NULL
)
}
\arguments{
\item{x}{dbSpatial object, SpatVector, or data.frame-like object with points coordinate
information (x, y, feat_ID)}

\item{feat_type}{character. feature type. Provide more than one value if
using the \code{split_keyword} param. For each set of keywords to split by, an
additional feat_type should be provided in the same order.}

\item{verbose}{be verbose}

\item{split_keyword}{list of character vectors of keywords to split the
giottoPoints object based on their feat_ID. Keywords will be \code{grepl()}
matched against the feature IDs information.}

\item{unique_IDs}{(optional) character vector of unique IDs present within
the spatVector data. Provided for cacheing purposes}
}
\value{
giottoPoints object wrapping a dbSpatial object
}
\description{
Create a \code{giottoPoints} object that wraps a dbSpatial points object
to support larger-than-memory spatial point data. This implementation extends
the standard GiottoClass implementation by providing specific methods for
dbSpatial objects.
}
\concept{Subcellular workflow}
