% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGiottoPolygon.R
\name{createGiottoPolygon}
\alias{createGiottoPolygon}
\alias{createGiottoPolygon,dbSpatial-method}
\title{Create GiottoPolygon object using dbSpatial}
\usage{
\S4method{createGiottoPolygon}{dbSpatial}(
  x,
  name = "polygons",
  verbose = TRUE,
  split_keyword = NULL,
  unique_IDs = NULL,
  calc_centroids = FALSE
)
}
\arguments{
\item{x}{dbSpatial object, SpatVector, or data.frame-like object with polygon coordinate
information (must include poly_ID column)}

\item{name}{character. Name for the polygon object}

\item{verbose}{be verbose}

\item{split_keyword}{list of character vectors of keywords to split the
giottoPolygon object based on their poly_ID. Keywords will be \code{grepl()}
matched against the polygon IDs information.}

\item{unique_IDs}{(optional) character vector of unique IDs present within
the spatVector data. Provided for cacheing purposes}

\item{calc_centroids}{logical. Whether to calculate centroids for the polygons}
}
\value{
giottoPolygon object wrapping a dbSpatial object
}
\description{
Create a \code{giottoPolygon} object that wraps a dbSpatial polygons object
to support larger-than-memory spatial polygon data. This implementation extends
the standard GiottoClass implementation by providing specific methods for
dbSpatial objects.
}
