% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dbBackend}
\alias{dbBackend}
\alias{dbBackend,missing-method}
\alias{dbBackend,backendInfo-method}
\alias{dbBackend,list-method}
\title{GiottoDB backend}
\usage{
\S4method{dbBackend}{missing}(
  dbdir = ":temp:",
  factory = duckdb::duckdb(),
  extension = ".duckdb",
  with_login = FALSE,
  verbose = TRUE,
  ...
)

\S4method{dbBackend}{backendInfo}(x, with_login = FALSE, verbose = TRUE, ...)

\S4method{dbBackend}{list}(x, with_login = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{dbdir}{directory to create a backend database}

\item{factory}{a function called to create database driver objects for the
backend. Default is `duckdb::duckdb()`}

\item{extension}{file extension (default = '.duckdb')}

\item{with_login}{(default = FALSE) whether a login is needed}

\item{verbose}{be verbose}

\item{...}{additional params to pass to pool::dbPool()}

\item{x}{database driver (default is duckdb::duckdb()), `backendInfo` object,
or or list of 1. DB driver call in string format, 2. path to the existing DB
file.}
}
\value{
invisibly returns backend ID
}
\description{
Defines and creates a database connection pool to be used by the backend
when a database driver function is provided.
When a `backendInfo` or `list` input is provided to `x` param instead, a
reconnection to an existing GiottoDB backend will be performed.
}
