% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dbBackendClose}
\alias{dbBackendClose}
\alias{dbBackendClose,gdbBackendID-method}
\alias{dbBackendClose,gdbBackend-method}
\alias{dbBackendClose,dbData-method}
\alias{dbBackendClose,environment-method}
\alias{dbBackendClose,missing-method}
\title{Disconnect backend from database}
\usage{
\S4method{dbBackendClose}{gdbBackendID}(x, ...)

\S4method{dbBackendClose}{gdbBackend}(x, ...)

\S4method{dbBackendClose}{dbData}(x, shutdown = TRUE)

\S4method{dbBackendClose}{environment}(x, ...)

\S4method{dbBackendClose}{missing}()
}
\arguments{
\item{x}{`gdbBackendID` of backend to close (optional) or `dbData`}
}
\value{
NULL invisibly
}
\description{
Closes pools. If specific backend_ID(s) are given then those will be closed.
When no specific ID is given, all existing backends will be closed.
If a `dbData` object is provided instead, closes the associated backend.
}
