% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dbBackendList}
\alias{dbBackendList}
\title{Existing GiottoDB backends info}
\usage{
dbBackendList()
}
\value{
data.table of backend IDs and database paths
}
\description{
Existing GiottoDB backends info
}
\examples{
dbBackend()
dbBackendList()
}
