% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dbIsValid}
\alias{dbIsValid}
\alias{dbIsValid,dbData-method}
\alias{dbIsValid,gdbBackendID-method}
\alias{dbIsValid,gdbBackend-method}
\title{Check if GiottoDB backend object has valid connection}
\usage{
\S4method{dbIsValid}{dbData}(dbObj)

\S4method{dbIsValid}{gdbBackendID}(dbObj)

\S4method{dbIsValid}{gdbBackend}(dbObj)
}
\arguments{
\item{dbObj}{GiottoDB object to check for database valid connection}
}
\description{
The method for `dbData` objects checks if they possess a pool that is valid.
The methods for `gdbBackend` or `gdbBackendID` check if the backend has a
valid connection. The two are not the same since if a `dbData` object does
not have a valid connection, it can still pull a valid connection from the
backend.
}
