% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{.as_mosaic_rows}
\alias{.as_mosaic_rows}
\title{Prepare rows for Mosaic inline JSON data}
\usage{
.as_mosaic_rows(data)
}
\arguments{
\item{data}{A data.frame to serialize into row objects}
}
\value{
List of row-wise named lists
}
\description{
Prepare rows for Mosaic inline JSON data
}
\keyword{internal}
