% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{.calc_cov_group_hvf_db}
\alias{.calc_cov_group_hvf_db}
\title{Calculate COV group HVF}
\usage{
.calc_cov_group_hvf_db(
  feat_in_cells_detected,
  nr_expression_groups = 20,
  zscore_threshold = 1,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL
)
}
\arguments{
\item{feat_in_cells_detected}{Data.table with expression statistics}

\item{nr_expression_groups}{Number of expression groups}

\item{zscore_threshold}{Z-score threshold}

\item{show_plot}{Show plot}

\item{return_plot}{Return plot}

\item{save_plot}{Save plot}
}
\value{
List with data.table of results and optional plot
}
\description{
Helper function to calculate HVF using COV groups method
}
\keyword{internal}
