% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{.calc_cov_loess_hvf_db}
\alias{.calc_cov_loess_hvf_db}
\title{Calculate COV loess HVF}
\usage{
.calc_cov_loess_hvf_db(
  feat_in_cells_detected,
  difference_in_cov = 0.1,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL
)
}
\arguments{
\item{feat_in_cells_detected}{Data.table with expression statistics}

\item{difference_in_cov}{Minimum difference in COV}

\item{show_plot}{Show plot}

\item{return_plot}{Return plot}

\item{save_plot}{Save plot}
}
\value{
List with data.table of results and optional plot
}
\description{
Helper function to calculate HVF using COV loess method
}
\keyword{internal}
