% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{.calc_expr_cov_stats_db}
\alias{.calc_expr_cov_stats_db}
\title{Calculate coefficient of variance statistics}
\usage{
.calc_expr_cov_stats_db(expr_values, expression_threshold)
}
\arguments{
\item{expr_values}{Expression matrix}

\item{expression_threshold}{Expression threshold}
}
\value{
Data.table with expression statistics including COV
}
\description{
Helper function to calculate COV stats
}
\keyword{internal}
