% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{.calc_var_hvf_db}
\alias{.calc_var_hvf_db}
\title{Calculate variance for highly variable features}
\usage{
.calc_var_hvf_db(
  scaled_matrix,
  var_threshold = 1.5,
  var_number = NULL,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL
)
}
\arguments{
\item{scaled_matrix}{Matrix with scaled values}

\item{var_threshold}{Variance threshold for features}

\item{var_number}{Number of top variance features}

\item{show_plot}{Show plot}

\item{return_plot}{Return ggplot object}

\item{save_plot}{Save plot}
}
\value{
List with data.table of results and optional plot
}
\description{
Helper function to calculate variance for scaled matrix
}
\keyword{internal}
