% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{.darken_mosaic_colors}
\alias{.darken_mosaic_colors}
\title{Darken CSS color strings}
\usage{
.darken_mosaic_colors(colors, factor = 0.7)
}
\arguments{
\item{colors}{Character vector of CSS colors}

\item{factor}{Multiplicative darkening factor in \link{0, 1}}
}
\value{
Character vector of darkened CSS colors
}
\description{
Darken CSS color strings
}
\keyword{internal}
