% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{.extract_duckdb_connection}
\alias{.extract_duckdb_connection}
\title{Extract DuckDB Connection from GiottoDB Object}
\usage{
.extract_duckdb_connection(gobject)
}
\arguments{
\item{gobject}{GiottoDB object}
}
\value{
DBI connection or NULL
}
\description{
Extract DuckDB Connection from GiottoDB Object
}
\keyword{internal}
