% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{.fetch_viz_data}
\alias{.fetch_viz_data}
\title{Fetch Visualization Data}
\usage{
.fetch_viz_data(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spat_loc_name = NULL,
  sdimx = "sdimx",
  sdimy = "sdimy",
  cell_color = NULL,
  select_cells = NULL,
  select_cell_groups = NULL
)
}
\arguments{
\item{gobject}{GiottoDB object}

\item{spat_unit}{Spatial unit}

\item{feat_type}{Feature type}

\item{spat_loc_name}{Spatial locations name}

\item{sdimx}{X dimension name}

\item{sdimy}{Y dimension name}

\item{cell_color}{Column to color by}

\item{select_cells}{Cell IDs to select}

\item{select_cell_groups}{Cell groups to select}
}
\value{
List with spatial_coords, cell_metadata, and polygon_data
}
\description{
Fetch Visualization Data
}
\keyword{internal}
