% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{.generate_color_palette}
\alias{.generate_color_palette}
\title{Generate Color Palette}
\usage{
.generate_color_palette(
  values,
  color_as_factor = TRUE,
  cell_color_code = NULL,
  cell_color_gradient = NULL
)
}
\arguments{
\item{values}{Vector of values to color}

\item{color_as_factor}{Treat as factor}

\item{cell_color_code}{Color codes}

\item{cell_color_gradient}{Color gradient}
}
\value{
Named vector of colors
}
\description{
Generate Color Palette
}
\keyword{internal}
