% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{.mosaic_bbox_domain}
\alias{.mosaic_bbox_domain}
\title{Build a lightweight GeoJSON bbox polygon for Mosaic projection fitting}
\usage{
.mosaic_bbox_domain(xmin, ymin, xmax, ymax)
}
\arguments{
\item{xmin}{Minimum x}

\item{ymin}{Minimum y}

\item{xmax}{Maximum x}

\item{ymax}{Maximum y}
}
\value{
A GeoJSON-like polygon list or \code{NULL}
}
\description{
Build a lightweight GeoJSON bbox polygon for Mosaic projection fitting
}
\keyword{internal}
