% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb.R
\name{duckdb_extension}
\alias{duckdb_extension}
\title{DuckDB extensions}
\usage{
duckdb_extension(p, extension = NULL)
}
\arguments{
\item{p}{pool or DBI connection object. If missing, will retrieve dataframe
of duckdb extensions information.}

\item{extension}{character. When `p` is also provided, will install and load
the designated extension.}
}
\value{
invisibly returns TRUE after an installation or load
}
\description{
DuckDB extensions
}
