% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{evaluate_conn}
\alias{evaluate_conn}
\alias{evaluate_conn,character-method}
\alias{evaluate_conn,Pool-method}
\alias{evaluate_conn,DBIConnection-method}
\title{Convert between DB handles and identifiers}
\usage{
\S4method{evaluate_conn}{character}(conn, mode = c("pool", "conn", "path", "id"), ...)

\S4method{evaluate_conn}{Pool}(conn, mode = c("pool", "conn", "path", "id"), ...)

\S4method{evaluate_conn}{DBIConnection}(conn, mode = c("pool", "conn", "path", "id"), ...)
}
\arguments{
\item{conn}{object to convert from}

\item{mode}{desired output type. One of pool, conn, path, or id}

\item{...}{additional params to pass}
}
\description{
GiottoDB refers to the database in several ways in different
situations.
\itemize{
  \item **path** (filepath) usually used whenever user input is part of the
workflow since that is the easiest for the user to have access to.
  \item **id** (backend hash ID) a shortened and unique way for functions to
reference a specific database backend.
  \item **pool** (pool of connection objects) a pool object that can be used
  to generate a database connected tbl.
  \item **conn** (DBI connection) a DBIConnection object that can be used
  to generate a database connected tbl. Used in situations where pool objects
  are not possible to be used, such as with temporary table generation.
}
Conversions between any of the four types are possible except for conn to
pool. These conversions can be performed using evaluate_conn.
}
