% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{gdbReadLog}
\alias{gdbReadLog}
\title{Read from the last generated log file}
\usage{
gdbReadLog(filepath)
}
\arguments{
\item{filepath}{character. filepath to log file. If omitted, tries to find the
last created log (will not work after a crash)}
}
\description{
Read from the last generated log file
}
