% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_factory.R
\name{generate_viz_s3_code}
\alias{generate_viz_s3_code}
\title{Generate R code for S3 generic and methods}
\usage{
generate_viz_s3_code(fun_name, pkg_name = "GiottoVisuals")
}
\arguments{
\item{fun_name}{Name of the function (string) e.g. "spatPlot2D"}

\item{pkg_name}{Package where the original function lives e.g. "GiottoVisuals"}
}
\value{
A character string containing the R code
}
\description{
Generate R code for S3 generic and methods
}
