% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiottoDB-class.R
\name{GiottoDB}
\alias{GiottoDB}
\title{Create a new GiottoDB object}
\usage{
GiottoDB(con, ...)
}
\arguments{
\item{con}{A \link{DBI} connection to a \link{duckdb} database}

\item{...}{Additional arguments passed to the giotto constructor}
}
\value{
A GiottoDB object
}
\description{
Create a new GiottoDB object, which is a database-backed
implementation of the Giotto object.
}
\examples{
\dontrun{
# Create a new GiottoDB object with a new database connection
library(GiottoDB)
library(duckdb)
library(DBI)

# Create a connection to a DuckDB database
con <- dbConnect(duckdb(), dbdir = ":memory:")
dbSpatial::loadSpatial(con) # Load spatial extension

# Create a new GiottoDB object
gobj_db <- GiottoDB(con = con)

# Don't forget to close the connection when done
dbDisconnect(con, shutdown = TRUE)
}
}
