% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{h5_array_write}
\alias{h5_array_write}
\title{Iteratively write chunked overlaps results to HDF5Matrix}
\usage{
h5_array_write(
  x,
  name = NULL,
  backend_type = "HDF5Array",
  filepath = NULL,
  row_names,
  col_names,
  as.sparse = TRUE,
  write_step = c(1000, 1000),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{poly_ID/feat_ID/counts DB tbl to write}

\item{backend_type}{character. Type of DelayedArray backend to use.
Default is "HDF5Array"}

\item{row_names}{expected row names}

\item{col_names}{expected column names}

\item{as.sparse}{Whether the data should be written as sparse or not to the
RealizationSink derivative to create. Not all realization backends support
this.}

\item{\dots}{additional params to pass}
}
\description{
Iteratively write chunked overlaps results to HDF5Matrix
}
\keyword{internal}
