% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{loadGDB}
\alias{loadGDB}
\title{Load a GiottoDB dbData object}
\usage{
loadGDB(file)
}
\arguments{
\item{file}{filepath}
}
\value{
loaded in `dbData` object
}
\description{
Loads a GiottoDB `dbData` object with `readRDS()`. Repairs
object connection to database backend.
}
