% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{loadGiotto}
\alias{loadGiotto}
\title{Load Giotto object}
\usage{
loadGiotto(path_to_folder, ...)
}
\arguments{
\item{path_to_folder}{path to folder where object was stored}

\item{...}{additional parameters passed to methods}
}
\description{
Function to load Giotto or GiottoDB objects
}
\concept{Saving and loading}
