% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{loadGiottoDB}
\alias{loadGiottoDB}
\title{Load GiottoDB object with explicit connection}
\usage{
loadGiottoDB(path_to_folder, con, ...)
}
\arguments{
\item{path_to_folder}{path to folder where object was stored}

\item{con}{DBI connection to use}

\item{...}{additional parameters passed to loadGiotto}
}
\description{
Helper function to load a Giotto object and convert it to GiottoDB
using an existing connection.
}
\concept{Saving and loading}
