% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_create}
\alias{log_create}
\title{Create a log file}
\usage{
log_create(filedir = tempdir())
}
\arguments{
\item{filedir}{character. Directory to create a logfile}
}
\description{
Creates a file called 'log.txt' at the specified directory. If no directory
is provided, it defaults to `tempdir()`. The path is additionally written to
the option 'gdb.last_logpath' and the open file connection to
'gdb.last_logconn'
}
\keyword{internal}
