% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_to_dbdir}
\alias{log_to_dbdir}
\title{Create log file in DB directory}
\usage{
log_to_dbdir(p)
}
\arguments{
\item{p}{pool of connections}
}
\description{
Create log file in DB directory
}
\keyword{internal}
