% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_write}
\alias{log_write}
\title{Write to log file}
\usage{
log_write(file_conn, x = "", main = "")
}
\arguments{
\item{file_conn}{a file connection}

\item{x}{character vector. Content to write}

\item{main}{character. Title to assign log entry}
}
\description{
Write to log file
}
\keyword{internal}
