% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapToMatrix.R
\name{overlapToMatrix}
\alias{overlapToMatrix}
\title{Convert dbSpatial overlaps to a matrix}
\usage{
\S4method{overlapToMatrix}{dbSpatial}(
  x,
  col_names = NULL,
  row_names = NULL,
  feat_count_column = NULL,
  count_info_column = NULL,
  output = c("dbSparseMatrix", "Matrix", "data.table"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{object containing overlaps info. Can be giotto object or SpatVector
points or data.table of overlaps generated from \code{calculateOverlap}}

\item{col_names, row_names}{character vector. (optional) Set of row and col
names that are expected to exist. This fixes the dimensions of the matrix
since the overlaps information does not directly report rows and cols where
no values were detected.}

\item{feat_count_column}{column with count information. If a
column called "count" is present in the feature points data, it will be
automatically selected.}

\item{count_info_column}{deprecated. Use \code{feat_count_column} instead.}

\item{output}{data format/class to return the results as. Default is "Matrix"}

\item{verbose}{be verbose}

\item{...}{additional params to pass to methods}
}
\value{
A dbSparseMatrix, sparse Matrix, or data.table-like result,
depending on \code{output}.
}
\description{
Convert dbSpatial overlap results into a dbMatrix, sparse Matrix, or
data.table-like output for downstream Giotto workflows.
}
\concept{Subcellular workflow}
