% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{reconnectBackend}
\alias{reconnectBackend}
\alias{reconnectBackend,list-method}
\title{Reconnect GiottoDB backend}
\usage{
\S4method{reconnectBackend}{backendInfo}(x, with_login = FALSE, verbose = TRUE, ...)

\S4method{reconnectBackend}{list}(x, with_login = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{backendInfo object or list of 1. DB driver call in string format,
2. path to the existing DB file.}

\item{with_login}{(default = FALSE) flag to check R environment variables}

\item{verbose}{be verbose}

\item{...}{additional params to pass
for login info and/or prompt for password}
}
\description{
Reconnect GiottoDB backend
}
