% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCA.R
\name{runPCA}
\alias{runPCA}
\alias{runPCA.GiottoDB}
\alias{runPCA.giotto}
\title{Run PCA on GiottoDB with \code{\link[dbMatrix:dbMatrix-class]{dbMatrix::dbMatrix}} support}
\usage{
runPCA(gobject, ...)

\method{runPCA}{GiottoDB}(gobject, ...)

\method{runPCA}{giotto}(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = "normalized",
  name = "pca",
  feats_to_use = "hvf",
  return_gobject = TRUE,
  ncp = 100,
  center = TRUE,
  scale_unit = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{A \code{\link[GiottoClass:giotto-class]{GiottoClass::giotto}} or \code{\link{GiottoDB}} object}

\item{...}{Additional arguments passed to underlying PCA methods}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use}

\item{name}{name of PCA dimension reduction}

\item{feats_to_use}{features to use for PCA}

\item{return_gobject}{whether to return giotto object}

\item{ncp}{number of principal components}

\item{center}{center data before PCA}

\item{scale_unit}{scale features before PCA}

\item{verbose}{verbosity}
}
\description{
S3 generic for PCA that dispatches to \code{dbMatrix::db_svd} when expression
data is stored as \code{\link[dbMatrix:dbMatrix-class]{dbMatrix::dbMatrix}}.
}
\concept{Dimension reduction}
