% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{saveGiotto}
\alias{saveGiotto}
\alias{saveGiotto.GiottoDB}
\alias{saveGiotto.default}
\title{Save Giotto object}
\usage{
saveGiotto(gobject, ...)

\method{saveGiotto}{GiottoDB}(
  gobject,
  foldername = "saveGiottoDir",
  dir = getwd(),
  method = c("RDS", "qs"),
  method_params = list(),
  overwrite = FALSE,
  export_image = TRUE,
  image_filetype = "PNG",
  include_feat_coord = TRUE,
  verbose = TRUE,
  ...
)

\method{saveGiotto}{default}(
  gobject,
  foldername = "saveGiottoDir",
  dir = getwd(),
  method = c("RDS", "qs"),
  method_params = list(),
  overwrite = FALSE,
  export_image = TRUE,
  image_filetype = "PNG",
  include_feat_coord = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto or GiottoDB object}

\item{...}{additional parameters passed to methods}

\item{foldername}{Folder name}

\item{dir}{Directory where to create the folder}

\item{method}{method to save main object}

\item{method_params}{additional method parameters for RDS or qs}

\item{overwrite}{Overwrite existing folders}

\item{export_image}{logical. Write out an image when saving giottoLargeImage}

\item{image_filetype}{the image filetype to use. Default is "PNG"}

\item{include_feat_coord}{logical. Whether to keep feature coordinates}

\item{verbose}{be verbose}
}
\value{
Creates a directory with GiottoDB object information including database files
}
\description{
Generic function to save Giotto or GiottoDB objects
}
\details{
This method extends GiottoClass::saveGiotto to handle
database-backed objects. The database file (.db) is moved to the save
directory, and connection information is updated to point to the new location.
Note: The original database file will no longer exist at its original location after saving.
}
\section{Methods (by class)}{
\itemize{
\item \code{saveGiotto(GiottoDB)}: Save a GiottoDB object

\item \code{saveGiotto(default)}: Default method - delegates to GiottoClass

}}
