% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_objects}
\alias{simulate_objects}
\alias{sim_duckdb}
\alias{sim_dbDataFrame}
\alias{sim_dbPointsProxy}
\alias{sim_dbPolygonProxy}
\title{Simulate GiottoDB Objects}
\usage{
sim_duckdb(data = datasets::iris, name = "test", p = NULL)

sim_dbDataFrame(data = NULL, name = "df_test", key = NA_character_)

sim_dbPointsProxy(data = NULL)

sim_dbPolygonProxy(data = NULL)
}
\arguments{
\item{data}{data to use (optional)}

\item{name}{remote name of simulated remote table (optional)}

\item{p}{pool object (optional)}

\item{key}{character. Column to use as key when indexing.}
}
\description{
Create simulated GiottoDB objects in memory from pre-prepared data. Useful
for testing purposes and examples.
}
\section{Functions}{
\itemize{
\item \code{sim_duckdb()}: Simulate a duckdb connection dplyr tbl_Pool in memory

\item \code{sim_dbDataFrame()}: Simulate a dbDataFrame in memory

\item \code{sim_dbPointsProxy()}: Simulate a dbPointsProxy in memory

\item \code{sim_dbPolygonProxy()}: Simulate a dbPolygonProxy in memory

}}
\keyword{internal}
