% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{stream_reader_fun}
\alias{stream_reader_fun}
\alias{stream_reader_gef_tx}
\title{Streamable reader functions for specific file formats}
\usage{
stream_reader_gef_tx(
  x,
  n = 4000L,
  i = 0L,
  bin_size = "bin100",
  output = c("DT", "SV")
)
}
\arguments{
\item{x}{reader input. Usually character filepath to file}

\item{n}{units (however defined in reader) to read per round}

\item{i}{'i'th round of reading (0 indexed)}

\item{bin_size}{.gef data bin size to use}

\item{output}{return as DT (data.table) or SV (terra SpatVector)}
}
\description{
Functions to access files and then read them in a chunkwise manner. Params
`x`, `n`, and `i` are required. Outputs from these functions should be in
formats that are usable by the downstream write_fun or of classes that
have pre-defined specific write functions (currently `SpatVector`)
}
\details{
Rules:\n
read_fun must not throw error when fewer are read than expected.\cr
read_fun must not throw error when 0 are found.
}
