% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{stream_to_db}
\alias{stream_to_db}
\alias{stream_to_db,Pool,character,SpatVector-method}
\alias{stream_to_db,ANY,character,data.frame-method}
\alias{stream_to_db,ANY,character,character-method}
\title{Stream data and write to DB}
\usage{
\S4method{stream_to_db}{Pool,character,SpatVector}(p, remote_name, x, ...)

\S4method{stream_to_db}{ANY,character,data.frame}(
  p,
  remote_name,
  x,
  custom_table_fields = NULL,
  pk = NULL,
  overwrite = FALSE,
  ...
)

\S4method{stream_to_db}{ANY,character,character}(
  p,
  remote_name,
  x,
  read_fun,
  n = 10L,
  stop_cond = function(x) nrow(x) == 0L,
  callback = NULL,
  write_fun = NULL,
  verbose = TRUE,
  log_to = NULL,
  report_n_chunks = 100L,
  ...
)
}
\arguments{
\item{p}{connection info or object (must be coercible to pool through
[evaluate_conn])}

\item{remote_name}{character. name of table in database to write to.}

\item{x}{information to stream}

\item{\dots}{additional params to pass to write_fun}

\item{custom_table_fields}{named character vector of columns and manually
assigned data types along with any other in-line constraints}

\item{pk}{character. Which column(s) to select as the primary key}

\item{overwrite}{whether to overwrite if table already exists (default = FALSE)}

\item{read_fun}{function. Controls how information should be read. Should
expose formal arguments 'x', 'n', and 'i'}

\item{n}{number of units (usually rows) to read per chunk}

\item{stop_cond}{function. Function to run on each chunk. Will stop the
chunk read looping when evaluates to TRUE.}

\item{callback}{(optional) function. Single param function that will take a
chunk as input, then modify it into the format expected by the `write_fun`
and db representation}

\item{write_fun}{(optional) function. Controls how the chunk is written to
the database. Default functions will automatically run for `data.table` and
terra `SpatVector` chunks.}

\item{verbose}{be verbose}

\item{log_to}{directory path. Create a logfile in this directory that logs
the progress of the chunked read. Any specified directory must exist. Setting
`NULL` logs to the same directory as the database}

\item{report_n_chunks}{if `verbose = TRUE`, report in the console every
n chunks that are processed.}

\item{...}{
  Arguments passed on to \code{\link[=createTableBE]{createTableBE}}
  \describe{
    \item{\code{fields_df}}{data frame from which to determine default data types}
  }}
}
\description{
Stream specific information into the database backend, whether for initial
data ingestion or for chunked processes.\cr
if custom_table_fields is provided and the table does not yet exist then
a specific DB table will be created with constraints. In all cases,
values will be fed to `dbWriteTable` with `append = TRUE` so that the values
will be appended, and if a specific table does not yet exist, it will be created.
}
\keyword{internal}
