% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{stream_writer_fun}
\alias{stream_writer_fun}
\alias{stream_writer_gef_tx}
\title{Streamable writer functions for specific file formats}
\usage{
stream_writer_gef_tx(p, remote_name, x, ...)
}
\arguments{
\item{x}{writer input. Usually data.table}

\item{\dots}{additional params to pass to [createTableBE]}
}
\description{
Functions to take partial reads of files and then write them to DB in a
chunkwise manner. Params `x`, `n`, and `i` are required. Outputs from these
functions should be in formats that are usable by classes
}
