% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_factory.R
\name{write_viz_autogen}
\alias{write_viz_autogen}
\title{Write generated visualization code to file}
\usage{
write_viz_autogen(output_file = "R/viz_autogen.R")
}
\arguments{
\item{output_file}{Path to output file}
}
\description{
Write generated visualization code to file
}
