% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{writeTableBE}
\alias{writeTableBE}
\title{Write a persistent table to the database}
\usage{
writeTableBE(p, remote_name, value, ...)
}
\arguments{
\item{p}{database connection pool object}

\item{remote_name}{name to assign the table within the database}

\item{value}{dataframe (or coercible) to write to backend}

\item{...}{additional params to pass to DBI::dbWriteTable()}
}
\description{
Write a persistent table to the database
}
