% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStatistics.R
\name{addStatistics}
\alias{addStatistics}
\title{Add feature and cell statistics for GiottoDB objects}
\usage{
addStatistics(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{...}{additional arguments passed to methods}
}
\value{
giotto object
}
\description{
S3 generic and method for adding statistics to GiottoDB objects
with database-optimized operations
}
