% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{calculateOverlap}
\alias{calculateOverlap}
\title{Calculate polygon overlapped features}
\usage{
\S4method{calculateOverlap}{dbSpatProxyData,dbSpatProxyData}(
  x,
  y,
  n_per_chunk = 1e+05,
  remote_name = result_count(),
  poly_subset_ids = NULL,
  feat_subset_column = NULL,
  feat_subset_ids = NULL,
  count_info_column = NULL,
  verbose = TRUE,
  ...
)
}
\description{
Calculate polygon overlapped features
}
