% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatProxy.R
\name{createDBPointsProxy}
\alias{createDBPointsProxy}
\title{Create dbPointsProxy object}
\usage{
createDBPointsProxy(
  SpatVector,
  remote_name = "pnt_test",
  db_path = ":temp:",
  xy_col = c("x", "y"),
  extent = NULL,
  overwrite = FALSE,
  chunk_size = 10000L,
  callback = NULL,
  custom_table_fields = NULL,
  ...
)
}
\arguments{
\item{SpatVector}{object coercible to SpatVector or filepath to spatial data
readable by \code{\link[terra]{vect}}}

\item{remote_name}{name of remote table on database backend}

\item{db_path}{filepath to the database backend}

\item{xy_col}{columns in data to read in that contain the x and y vertex info}

\item{extent}{terra SpatExtent (optional) that can be used to subset the data
to read in before it is saved to database}

\item{overwrite}{whether to overwrite if \code{remote_name} already exists on
the database}

\item{chunk_size}{the number of polygons to read in per chunk read}

\item{callback}{data formatting and manipulations to perform chunkwise before the data
is saved to database. Instructions should be provided as a function that takes
an input of a data.table and returns a data.table}

\item{custom_table_fields}{(optional) custom table field SQL settings to use
during table creation for the spatial geometry table}

\item{\dots}{addtional params to pass}
}
\description{
Create dbPointsProxy object
}
