% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiottoDB-class.R
\name{dbReconnect,GiottoDB-method}
\alias{dbReconnect,GiottoDB-method}
\title{Reconnect a GiottoDB object}
\usage{
\S4method{dbReconnect}{GiottoDB}(x)
}
\arguments{
\item{x}{A \code{\link{GiottoDB}} object.}
}
\value{
A \code{\link{GiottoDB}} object with a valid top-level database connection.
}
\description{
Reconnect a GiottoDB object
}
\concept{Connection management}
