% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{loadGDB}
\alias{loadGDB}
\title{Load a GiottoDB dbData object}
\usage{
loadGDB(file, silent = TRUE)
}
\arguments{
\item{file}{filepath}

\item{silent}{silences read-in console prints}
}
\value{
loaded in `dbData` object
}
\description{
Loads a GiottoDB `dbData` object with `readRDS()`. Also performs
initialization to repair the database connection.
}
\details{
Some warnings and messages are expected when loading GiottoDB `dbData`
objects due to the way that pool and dplyr work. These are diverted to a
temporary file with `sink()` if `silent` is TRUE. These messages should
have no effect on function.
}
