% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{overlapToMatrix}
\alias{overlapToMatrix}
\title{Create counts matrix from overlaps information}
\usage{
\S4method{overlapToMatrix}{dbPointsProxy}(
  x,
  remote_name = paste0(remoteName(x), "_raw"),
  gdb_name = NULL,
  col_names = NULL,
  row_names = NULL,
  count_info_column = NULL,
  output = c("HDF5Array"),
  write_step = c(1000, 1000),
  filepath = NULL,
  ...
)
}
\arguments{
\item{x}{object containing overlaps information}

\item{remote_name}{character. name to be applied to on-disk representation of
the results (if applicable)}

\item{gdb_name}{alias for remote_name param. For disambiguation in external
code referencing this method.}

\item{col_names, row_names}{expected col and row names,}

\item{count_info_column}{character. column containing counts expected for
each overlap found. Assumed to be 1 for all, if not provided.}

\item{write_step}{if output is HDF5Array, the dimensions of the chunks
that are written iteratively to disk. Default is c(1000,1000). Note that
actual write step used will respect the expected dimensions.}

\item{filepath}{character. Save directory for the on-disk representation of
the results (if applicable)}

\item{\dots}{additional params to pass}
}
\description{
Create counts matrix from overlaps information
}
