% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{stream_reader_fun}
\alias{stream_reader_fun}
\alias{stream_reader_fread}
\alias{stream_reader_arrow}
\alias{stream_reader_gef_tx}
\title{Streamable reader functions for specific file formats}
\usage{
stream_reader_fread(x, n, i, header = NULL, ...)

stream_reader_arrow(x, n, i, delim = ",", ...)

stream_reader_gef_tx(
  x,
  n = 4000L,
  i = 0L,
  bin_size = "bin100",
  output = c("DT", "SV")
)
}
\arguments{
\item{x}{reader input. Usually character filepath to file}

\item{n}{units (however defined in reader) to read per round}

\item{i}{'i'th round of reading (0 indexed)}

\item{header}{Whether the file has a header with colnames. Should never be
set to TRUE. Either leave as default or set FALSE when no colnames line
is present.}

\item{...}{additional params to pass to the specific reading function}

\item{delim}{delimiting character. Default is "," (csv)}

\item{bin_size}{.gef bin size to use (default is "bin100")}

\item{output}{return as DT (data.table) or SV (terra SpatVector)}
}
\description{
Functions to access files and then read them in a chunkwise manner. Params
`x`, `n`, and `i` are required. Outputs from these functions should be in
formats that are usable by the downstream write_fun or of classes that
have pre-defined specific write functions (currently `SpatVector`)
}
\details{
Rules:\cr
read_fun must not throw error when fewer are read than expected.\cr
read_fun must not throw error when 0 are found.
}
\section{Functions}{
\itemize{
\item \code{stream_reader_fread()}: data.table `fread()` chunked reads

\item \code{stream_reader_arrow()}: arrow chunked reads for delimited files

\item \code{stream_reader_gef_tx()}: chunked reads of .gef/.bgef stereoseq files
while treating them as spatial points/transcripts info

}}
